redefines get_job_models due to circular imports
