+++ import_export_celery/admin_actions.py
@@ -6,8 +6,7 @@ from django.utils.translation import gettext_lazy as _
 from django.urls import reverse
 from django.shortcuts import redirect
 
-from . import tasks
-from .models.utils import get_job_models
+from . import tasks, get_job_models
 
 
 ImportJob, ExportJob = get_job_models()
